Phase 2 client: tempo, metronome, bar ruler, syllable anchoring, playback
Tempo control (BPM, meter, tap tempo) in the song header. A raw Web
Audio lookahead scheduler (web/src/audio/metronome.ts) ticks at
sixteenth-note resolution so playback highlighting can track syllables
between beats, sounding a click only on the beat with an accented
downbeat.

Bar ruler per section, sized from bar_count/meter, laid out as a single
timeline across the song (each section starts right after the bars
consumed by the ones before it). Lines assign to a bar by dragging onto
the ruler or typing start_bar/start_beat directly.

Syllable anchoring reuses the existing segmentation strip: clicking a
syllable pins it (reusing --pulse for the visual state, no new accent),
clicking again unpins it. The client never reimplements the resolver's
interpolation — pinning reconstructs a line's full anchor list from the
grid response's resolved bar/beat (exact for pinned syllables, since
those are resolveLineTiming()'s fixed points) and PUTs the whole set.

Also fixes a pre-existing Dockerfile bug: postinstall's cmudict fetch
script wasn't copied in before `npm ci`, so every server-stage build
failed with MODULE_NOT_FOUND. Both stages now skip lifecycle scripts,
since the build stage never needs the dictionary and the runtime stage
gets it from the vendored data/ copy instead.

+++ src/types.ts
@@ -50,7 +50,7 @@ export const RevisionCreate = z.object({ label: z.string().max(120).default('') 
 /** offset is sixteenth notes from the line's start; index is ordinal into analyzeLine().words.flatMap(syllables). */
 export const SyllableAnchor = z.object({
   index: z.number().int().min(0),
-  offset: z.number().min(0),
+  offset: z.number().int().min(0),
 });
 
 export const LineTimingBody = z.object({
